Add option to get undirected graphs when merging edges.
This functionality is needed for the structural balance algorithm,
which takes as input an undirected graph with signed edges. Since
we have many parallel edges in the network we are working with,
their sentiment needs to be condensated into a sign, either positive
or negative. If the sign is ambiguous, the edge is deleted.